refactor(codegen): move GenerateProvidersRegistry to codegen internal package
